Update and rename ASY4.sh to ASY5.sh

+++ ASY5.sh
@@ -3,16 +3,22 @@
 ZAPRET_CONF="/etc/config/zapret"
 STR_URL="https://raw.githubusercontent.com/StressOzz/Test/refs/heads/main/ListStrYou"
 TMP_LIST="/tmp/zapret_yt_list.txt"
+SAVED_STR="/opt/StrYou"
 
 TEST_HOST="https://rr1---sn-gvnuxaxjvh-jx3z.googlevideo.com"
 TIMEOUT=5
 WAIT_AFTER_APPLY=3
-RESULT_FILE="/tmp/zapret_strategy_found"
+
+# Проверяем, есть ли сохранённая стратегия
+if [ -f "$SAVED_STR" ]; then
+    STRATEGY_NAME=$(head -n1 "$SAVED_STR")
+    echo "[ZAPRET] Используем сохранённую стратегию: $STRATEGY_NAME"
+    exit 0  # или можно сразу применить, если нужно
+fi
 
 # Скачать список стратегий
 curl -fsSL "$STR_URL" -o "$TMP_LIST" || { echo "Не удалось скачать список"; exit 1; }
 
-# Посчитать количество стратегий
 TOTAL=$(grep -c '^YT[0-9]\+' "$TMP_LIST")
 echo "[ZAPRET] Найдено стратегий: $TOTAL"
 echo
@@ -21,7 +27,6 @@ CURRENT_NAME=""
 CURRENT_BODY=""
 COUNT=0
 
-# Прогресс-бар
 progress_bar() {
     done="$1"
     total="$2"
@@ -33,7 +38,6 @@ progress_bar() {
     echo "[$BAR] $done/$total"
 }
 
-# Применение стратегии
 apply_strategy() {
     NAME="$1"
     BODY="$2"
@@ -44,10 +48,9 @@ apply_strategy() {
         printf "%b\n" "$BODY"
         echo "'"
     } >> "$ZAPRET_CONF"
-    chmod +x /opt/zapret/sync_config.sh; /opt/zapret/sync_config.sh; /etc/init.d/zapret restart >/dev/null 2>&1;
+    /etc/init.d/zapret restart >/dev/null 2>&1
 }
 
-# Проверка доступа
 check_access() {
     curl -I -s --connect-timeout "$TIMEOUT" -m "$TIMEOUT" -o /dev/null -w "%{http_code}" "$TEST_HOST"
 }
@@ -68,8 +71,8 @@ while IFS= read -r LINE || [ -n "$LINE" ]; do
                 echo "Enter — оставить стратегию, N — продолжить перебор"
                 read -r ANSWER </dev/tty
                 if [ -z "$ANSWER" ]; then
-                    echo "🏁 Рабочая стратегия: $CURRENT_NAME"
-                    echo "$CURRENT_NAME" > "$RESULT_FILE"
+                    echo "$CURRENT_NAME" > "$SAVED_STR"
+                    echo "🏁 Рабочая стратегия: $CURRENT_NAME сохранена в $SAVED_STR"
                     exit 0
                 fi
             else
@@ -98,8 +101,8 @@ if [ -n "$CURRENT_NAME" ]; then
         echo "Enter — оставить стратегию, N — продолжить перебор"
         read -r ANSWER </dev/tty
         if [ -z "$ANSWER" ]; then
-            echo "🏁 Рабочая стратегия: $CURRENT_NAME"
-            echo "$CURRENT_NAME" > "$RESULT_FILE"
+            echo "$CURRENT_NAME" > "$SAVED_STR"
+            echo "🏁 Рабочая стратегия: $CURRENT_NAME сохранена в $SAVED_STR"
             exit 0
         fi
     else
